fix: rename Enrichment plugins to Before
